% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nugget-rules.R, R/as_nugget.R
\name{as_nugget.rules}
\alias{as_nugget.rules}
\alias{as_nugget}
\title{Convert arules::rules to nugget}
\usage{
\method{as_nugget}{rules}(x, ...)

as_nugget(x, ...)
}
\arguments{
\item{x}{An object of class \code{rules} from the \pkg{arules} package.}

\item{...}{Additional arguments (not used).}
}
\value{
A nugget object containing the rules and associated statistics.
}
\description{
Convert an object of class \code{rules} from the \pkg{arules} package to
a nugget.
}
\details{
The \code{rules} class is used in the \pkg{arules} package to represent
association rules. It contains association rules along with various statistics
such as support, confidence, and lift. Object of class \code{rules} can be
created using the \code{\link[arules:apriori]{arules::apriori()}} or \code{\link[arules:eclat]{arules::eclat()}} functions
and \code{\link[arules:ruleInduction]{arules::ruleInduction()}} function.

The resulting nugget object can be used for further analysis and exploration
using the functions provided by the \pkg{nuggets} package.
}
\examples{
# Prepare data
d <- partition(mtcars, .breaks = 2)

# Use arules::apriori() to search for association rules
fit1 <- arules::apriori(d,
                        parameter = list(minlen = 1,
                                         maxlen = 6,
                                         supp = 0.001,
                                         conf = 0.5))

# Convert the rules to a nugget
rules1 <- as_nugget(fit1)

# Test whether we have the nugget
is_nugget(rules1)

# Explore the results from arules interactively
\dontrun{
explore(rules1)
}

# Use arules::eclat() to search for association rules
fit2 <- arules::eclat(d,
                      parameter = list(minlen = 1,
                                       maxlen = 6,
                                       supp=0.001))
fit2 <- arules::ruleInduction(fit2, confidence = 0.5)

# Convert the rules to a nugget
rules2 <- as_nugget(fit2)

# Test whether we have the nugget
is_nugget(rules2)

# Explore the results from arules interactively
\dontrun{
explore(rules2)
}
}
