% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_ancestors-associations.R, R/dig_ancestors.R
\name{dig_ancestors.associations}
\alias{dig_ancestors.associations}
\alias{dig_ancestors}
\title{Dig ancestors of an association rule}
\usage{
\method{dig_ancestors}{associations}(x, data, ...)

dig_ancestors(x, data, ...)
}
\arguments{
\item{x}{A nugget of flavour \code{associations} containing a single association rule.
(Typically created with \code{\link[=dig_associations]{dig_associations()}}.)}

\item{data}{a matrix or data frame with data to search in. The matrix must be
numeric (double) or logical. If \code{x} is a data frame then each column
must be either numeric (double) or logical.}

\item{...}{further arguments (currently not used).}
}
\value{
A nugget of flavour \code{associations} containing all association rules that are
ancestors of the given rule \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Searches for all association rules that are ancestors of the given association
rule, i.e. all rules whose antecedent is a subset of the antecedent of the
given rule and whose consequent is equal to the consequent of the given rule.
The search is performed using the same \code{disjoint}, \code{excluded}, and \code{t_norm}
parameters as the original search that produced the given rule.
}
\examples{
d <- partition(mtcars, .breaks = 2)
rules <- dig_associations(d,
                          antecedent = !starts_with("mpg"),
                          consequent = starts_with("mpg"),
                          min_support = 0.3,
                          min_confidence = 0.8)
r <- rules[1, ]  # get first rule
anc <- dig_ancestors(r, d)
}
\seealso{
\code{\link[=dig_associations]{dig_associations()}}
}
\author{
Michal Burda
}
