% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_contrasts.R
\name{dig_contrasts}
\alias{dig_contrasts}
\title{Search for contrast patterns}
\usage{
dig_contrasts(
  x,
  condition = where(is.logical),
  xvars = where(is.numeric),
  yvars = where(is.numeric),
  method = "t",
  alternative = "two.sided",
  min_length = 0L,
  max_length = Inf,
  min_support = 0,
  max_p_value = 0.05,
  threads = 1,
  ...
)
}
\arguments{
\item{x}{a matrix or data frame with data to search in.}

\item{condition}{a tidyselect expression (see
\href{https://tidyselect.r-lib.org/articles/syntax.html}{tidyselect syntax})
specifying the columns to use as condition predicates}

\item{xvars}{a tidyselect expression (see
\href{https://tidyselect.r-lib.org/articles/syntax.html}{tidyselect syntax})
specifying the columns to use for computation of contrasts}

\item{yvars}{a tidyselect expression (see
\href{https://tidyselect.r-lib.org/articles/syntax.html}{tidyselect syntax})
specifying the columns to use for computation of contrasts}

\item{method}{a character string indicating which contrast to compute.
One of \code{"t"}, \code{"wilcox"}, or \code{"var"}. \code{"t"} (resp. \code{"wilcos"}) compute
a parametric (resp. non-parametric) test on equality in position, and
\code{"var"} performs the F-test on equality of variance.}

\item{alternative}{indicates the alternative hypothesis and must be one of
\code{"two.sided"}, \code{"greater"} or \code{"less"}. \code{"greater"} corresponds to
positive association, \code{"less"} to negative association.}

\item{min_length}{the minimum size (the minimum number of predicates) of the
condition to be generated (must be greater or equal to 0). If 0, the empty
condition is generated in the first place.}

\item{max_length}{The maximum size (the maximum number of predicates) of the
condition to be generated. If equal to Inf, the maximum length of conditions
is limited only by the number of available predicates.}

\item{min_support}{the minimum support of a condition to trigger the callback
function for it. The support of the condition is the relative frequency
of the condition in the dataset \code{x}. For logical data, it equals to the
relative frequency of rows such that all condition predicates are TRUE on it.
For numerical (double) input, the support is computed as the mean (over all
rows) of multiplications of predicate values.}

\item{max_p_value}{the maximum p-value of a test for the pattern to be considered
significant. If the p-value of the test is greater than \code{max_p_value}, the
pattern is not included in the result.}

\item{threads}{the number of threads to use for parallel computation.}

\item{...}{Further arguments passed to the underlying test function
(\code{\link[=t.test]{t.test()}}, \code{\link[=wilcox.test]{wilcox.test()}}, or \code{\link[=var.test]{var.test()}} accordingly to the
selected method).}
}
\value{
A tibble with found rules.
}
\description{
Contrast patterns are a generalization of association rules that allow
for the specification of a condition under which there is a significant
difference between two numeric variables. The contrast is computed using
a statistical test, which is specified by the \code{method} argument. The
function computes the contrast between all pairs of variables, where the
first variable is specified by the \code{xvars} argument and the second variable
is specified by the \code{yvars} argument. The contrast is computed in sub-data
corresponding to conditions generated from the \code{condition} columns.
}
\examples{
crispCO2 <- partition(CO2, Plant:Treatment)
dig_contrasts(crispCO2,
              condition = where(is.logical),
              xvars = conc,
              yvars = uptake,
              method = "t",
              min_support = 0.1)

}
\seealso{
\code{\link[=dig]{dig()}}, \code{\link[=dig_grid]{dig_grid()}}, \code{\link[stats:t.test]{stats::t.test()}}, \code{\link[stats:wilcox.test]{stats::wilcox.test()}}, \code{\link[stats:var.test]{stats::var.test()}}
}
\author{
Michal Burda
}
