% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_itemsets.R
\name{dig_itemsets}
\alias{dig_itemsets}
\title{Search for itemsets}
\usage{
dig_itemsets(
  x,
  items = everything(),
  weights = NULL,
  disjoint = var_names(colnames(x)),
  excluded = NULL,
  min_length = 0L,
  max_length = Inf,
  min_support = 0,
  t_norm = "goguen",
  max_results = Inf,
  verbose = FALSE,
  error_context = list(arg_x = "x", arg_items = "items", arg_weights = "weights",
    arg_disjoint = "disjoint", arg_excluded = "excluded", arg_min_length = "min_length",
    arg_max_length = "max_length", arg_min_support = "min_support", arg_t_norm =
    "t_norm", arg_max_results = "max_results", arg_verbose = "verbose", arg_threads =
    "threads", call = current_env())
)
}
\arguments{
\item{x}{a matrix or data frame with data to search in. The matrix must be
numeric (double) or logical. If \code{x} is a data frame then each column
must be either numeric (double) or logical.}

\item{items}{a tidyselect expression (see
\href{https://tidyselect.r-lib.org/articles/syntax.html}{tidyselect syntax})
specifying the columns to use as item candidates.}

\item{weights}{\code{NULL} or a tidyselect expression (see
\href{https://tidyselect.r-lib.org/articles/syntax.html}{tidyselect syntax})
selecting one numeric, finite, non-negative column of observation
weights.}

\item{disjoint}{an atomic vector of size equal to the number of columns of \code{x}
that specifies the groups of predicates: if some elements of the \code{disjoint}
vector are equal, then the corresponding columns of \code{x} will NOT be
present together in a single condition. If \code{x} is prepared with
\code{\link[=partition]{partition()}}, using the \code{\link[=var_names]{var_names()}} function on \code{x}'s column names
is a convenient way to create the \code{disjoint} vector.}

\item{excluded}{\code{NULL} or a list of character vectors, each representing a
known implication (axiom). In each vector, all but the last element
form the antecedent and the last element is the consequent.}

\item{min_length}{the minimum length, i.e., the minimum number of predicates
in an itemset. Value must be greater or equal to 0.}

\item{max_length}{The maximum length, i.e., the maximum number of predicates
in an itemset. If equal to Inf, the maximum length is limited only by the
number of available predicates.}

\item{min_support}{the minimum support of an itemset in the dataset \code{x}.}

\item{t_norm}{a t-norm used to compute conjunction of truth degrees. It must be one of
\code{"goedel"} (minimum t-norm), \code{"goguen"} (product t-norm), or \code{"lukas"}
(Łukasiewicz t-norm).}

\item{max_results}{the maximum number of generated itemsets. If the number of
found itemsets exceeds \code{max_results}, only the first \code{max_results}
itemsets are returned.}

\item{verbose}{a logical value indicating whether to print progress messages.}

\item{error_context}{a named list providing context for error messages.}
}
\value{
An S3 object, which is an instance of \code{itemsets} and \code{nugget}
classes, and which is a tibble with found itemsets and basic quality
measures.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{dig_itemsets()} searches for itemsets (sets of predicates) in \code{x}.
The support of an itemset is the relative frequency of rows satisfying all
predicates in the set.
}
\examples{
d <- partition(mtcars, .breaks = 2)
dig_itemsets(d, items = everything(), min_support = 0.3)
}
\seealso{
\code{\link[=partition]{partition()}}, \code{\link[=var_names]{var_names()}}, \code{\link[=dig]{dig()}}
}
