% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-associations.R
\name{explore.associations}
\alias{explore.associations}
\title{Show interactive application to explore association rules}
\usage{
\method{explore}{associations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{associations}, typically created with
\code{\link[=dig_associations]{dig_associations()}}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
If printed in an interactive R session, the application will be run
immediately.
}
\description{
Show interactive application to explore association rules
}
\author{
Michal Burda
}
