% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-baseline_contrasts.R
\name{explore.baseline_contrasts}
\alias{explore.baseline_contrasts}
\title{Show interactive application to explore baseline contrasts}
\usage{
\method{explore}{baseline_contrasts}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of S3 class \code{baseline_contrasts}, typically created with
\code{\link[=dig_baseline_contrasts]{dig_baseline_contrasts()}}.}

\item{data}{An optional data frame containing the dataset from which the
contrasts were computed. Providing this enables additional contextual
features in the explorer, such as examining supporting records.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
When "printed" in an interactive R session, the application is launched
immediately in the default web browser.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of baseline
contrast patterns. The explorer provides tools for inspecting pattern quality,
comparing measures, and interactively filtering subsets of patterns.
When the original dataset is supplied, the application also allows
for contextual exploration of contrasts with respect to the underlying data.
}
\examples{
\dontrun{
d <- partition(mtcars, .breaks = 2, .keep = TRUE)
res <- dig_baseline_contrasts(d,
                              condition = where(is.logical),
                              vars = where(is.numeric),
                              min_support = 0.3,
                              max_length = 2)

# launch the interactive explorer
explore(res, data = d)
}
}
\seealso{
\code{\link[=dig_baseline_contrasts]{dig_baseline_contrasts()}}
}
\author{
Michal Burda
}
