% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-correlations.R
\name{explore.correlations}
\alias{explore.correlations}
\title{Show interactive application to explore conditional correlations}
\usage{
\method{explore}{correlations}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of S3 class \code{correlations}, typically created with
\code{\link[=dig_correlations]{dig_correlations()}}.}

\item{data}{An optional data frame containing the dataset from which the
correlations were computed. Providing this enables additional contextual
features in the explorer, such as examining supporting records.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
When "printed" in an interactive R session, the application is launched
immediately in the default web browser.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of
conditional correlation patterns. The explorer provides tools for inspecting
pattern quality, comparing measures, and interactively filtering subsets
of patterns. When the original dataset is supplied, the application also
allows for contextual exploration of correlations with respect to the
underlying data.
}
\examples{
\dontrun{
d <- partition(iris, Species)
res <- dig_correlations(d,
                        condition = where(is.logical),
                        xvars = Sepal.Length:Petal.Width,
                        yvars = Sepal.Length:Petal.Width)

# launch the interactive explorer
explore(res, data = d)
}
}
\seealso{
\code{\link[=dig_correlations]{dig_correlations()}}
}
\author{
Michal Burda
}
