% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-itemsets.R
\name{explore.itemsets}
\alias{explore.itemsets}
\title{Show interactive application to explore itemsets}
\usage{
\method{explore}{itemsets}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of S3 class \code{itemsets}, typically created with
\code{\link[=dig_itemsets]{dig_itemsets()}}.}

\item{data}{An optional data frame containing the dataset from which the
itemsets were computed. Currently unused.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
When "printed" in an interactive R session, the application is launched
immediately in the default web browser.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of itemsets.
The explorer provides tools for inspecting pattern quality,
comparing measures, and interactively filtering subsets of patterns.
}
\examples{
\dontrun{
d <- partition(mtcars, .breaks = 2, .keep = TRUE)
res <- dig_itemsets(d,
                    items = where(is.logical),
                    min_support = 0.3,
                    max_length = 2)

# launch the interactive explorer
explore(res)
}
}
\seealso{
\code{\link[=dig_itemsets]{dig_itemsets()}}
}
\author{
Michal Burda
}
