% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-paired_baseline_contrasts.R
\name{explore.paired_baseline_contrasts}
\alias{explore.paired_baseline_contrasts}
\title{Show interactive application to explore paired baseline contrasts}
\usage{
\method{explore}{paired_baseline_contrasts}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of S3 class \code{paired_baseline_contrasts}, typically
created with \code{\link[=dig_paired_baseline_contrasts]{dig_paired_baseline_contrasts()}}.}

\item{data}{An optional data frame containing the dataset from which the
contrasts were computed. Providing this enables additional contextual
features in the explorer, such as examining supporting records.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
When "printed" in an interactive R session, the application is launched
immediately in the default web browser.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of paired
baseline contrast patterns. The explorer provides tools for inspecting
pattern quality, comparing measures, and interactively filtering subsets
of patterns. When the original dataset is supplied, the application also
allows for contextual exploration of contrasts with respect to the
underlying data.
}
\examples{
\dontrun{
crispIris <- iris
crispIris$Sepal.Ratio <- iris$Sepal.Length / iris$Sepal.Width
crispIris$Petal.Ratio <- iris$Petal.Length / iris$Petal.Width
crispIris <- partition(crispIris, Species)
res <- dig_paired_baseline_contrasts(crispIris,
                                     condition = where(is.logical),
                                     xvars = Sepal.Ratio,
                                     yvars = Petal.Ratio,
                                     method = "t",
                                     min_support = 0.1)

# launch the interactive explorer
explore(res, data = crispIris)
}
}
\seealso{
\code{\link[=dig_paired_baseline_contrasts]{dig_paired_baseline_contrasts()}}
}
\author{
Michal Burda
}
