% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-associations.R,
%   R/explore-baseline_contrasts.R, R/explore-complement_contrasts.R,
%   R/explore-correlations.R, R/explore-paired_baseline_contrasts.R
\name{explore.associations}
\alias{explore.associations}
\alias{explore.baseline_contrasts}
\alias{explore.complement_contrasts}
\alias{explore.correlations}
\alias{explore.paired_baseline_contrasts}
\title{Show interactive application to explore association rules}
\usage{
\method{explore}{associations}(x, data = NULL, ...)

\method{explore}{baseline_contrasts}(x, data = NULL, ...)

\method{explore}{complement_contrasts}(x, data = NULL, ...)

\method{explore}{correlations}(x, data = NULL, ...)

\method{explore}{paired_baseline_contrasts}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object of S3 class \code{associations}, typically created with
\code{\link[=dig_associations]{dig_associations()}}.}

\item{data}{An optional data frame containing the dataset from which the
rules were mined. Providing this enables additional contextual features in
the explorer, such as examining supporting records.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{shiny.appobj} representing the Shiny application.
When "printed" in an interactive R session, the application is launched
immediately in the default web browser.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of mined
association rules. The explorer provides tools for inspecting rule quality,
comparing interestingness measures, and interactively filtering subsets of
rules. When the original dataset is supplied, the application also allows
for contextual exploration of rules with respect to the underlying data.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of baseline
contrast patterns. The explorer provides tools for inspecting pattern quality,
comparing measures, and interactively filtering subsets of patterns.
When the original dataset is supplied, the application also allows
for contextual exploration of contrasts with respect to the underlying data.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of
complement contrast patterns. The explorer provides tools for inspecting
pattern quality, comparing measures, and interactively filtering subsets
of patterns. When the original dataset is supplied, the application also
allows for contextual exploration of contrasts with respect to the
underlying data.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of
conditional correlation patterns. The explorer provides tools for inspecting
pattern quality, comparing measures, and interactively filtering subsets
of patterns. When the original dataset is supplied, the application also
allows for contextual exploration of correlations with respect to the
underlying data.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Launches an interactive Shiny application for visual exploration of paired
baseline contrast patterns. The explorer provides tools for inspecting
pattern quality, comparing measures, and interactively filtering subsets
of patterns. When the original dataset is supplied, the application also
allows for contextual exploration of contrasts with respect to the
underlying data.
}
\examples{
\dontrun{
data("iris")
# convert all columns into dummy logical variables
part <- partition(iris, .breaks = 3)

# find association rules
rules <- dig_associations(part)

# launch the interactive explorer
explore(rules, data = part)
}
\dontrun{
d <- partition(mtcars, .breaks = 2, .keep = TRUE)
res <- dig_baseline_contrasts(d,
                              condition = where(is.logical),
                              vars = where(is.numeric),
                              min_support = 0.3,
                              max_length = 2)

# launch the interactive explorer
explore(res, data = d)
}
\dontrun{
d <- partition(mtcars, .breaks = 2, .keep = TRUE)
res <- dig_complement_contrasts(d,
                                condition = where(is.logical),
                                vars = where(is.numeric),
                                min_support = 0.3,
                                max_length = 2)

# launch the interactive explorer
explore(res, data = d)
}
\dontrun{
d <- partition(iris, Species)
res <- dig_correlations(d,
                        condition = where(is.logical),
                        xvars = Sepal.Length:Petal.Width,
                        yvars = Sepal.Length:Petal.Width)

# launch the interactive explorer
explore(res, data = d)
}
\dontrun{
crispIris <- iris
crispIris$Sepal.Ratio <- iris$Sepal.Length / iris$Sepal.Width
crispIris$Petal.Ratio <- iris$Petal.Length / iris$Petal.Width
crispIris <- partition(crispIris, Species)
res <- dig_paired_baseline_contrasts(crispIris,
                                     condition = where(is.logical),
                                     xvars = Sepal.Ratio,
                                     yvars = Petal.Ratio,
                                     method = "t",
                                     min_support = 0.1)

# launch the interactive explorer
explore(res, data = crispIris)
}
}
\seealso{
\code{\link[=dig_associations]{dig_associations()}}

\code{\link[=dig_baseline_contrasts]{dig_baseline_contrasts()}}

\code{\link[=dig_complement_contrasts]{dig_complement_contrasts()}}

\code{\link[=dig_correlations]{dig_correlations()}}

\code{\link[=dig_paired_baseline_contrasts]{dig_paired_baseline_contrasts()}}
}
\author{
Michal Burda
}
