% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\name{fire}
\alias{fire}
\title{Obtain truth-degrees of conditions}
\usage{
fire(x, condition, t_norm = "goguen")
}
\arguments{
\item{x}{a matrix or data frame. The matrix must be numeric (double) or logical.
If \code{x} is a data frame then each column must be either numeric (double) or
logical.}

\item{condition}{a character vector of conditions, each element as formatted
by \code{\link[=format_condition]{format_condition()}}. E.g., \code{"{p1,p2,p3}"} is a condition with three
predicates \code{"p1"}, \code{"p2"}, and \code{"p3"}. All predicates present in the
condition must exist as column names in \code{x}.}

\item{t_norm}{a t-norm used to compute conjunction of weights. It must be one of
\code{"goedel"} (minimum t-norm), \code{"goguen"} (product t-norm), or \code{"lukas"}
(Lukasiewicz t-norm).}
}
\value{
A numeric matrix with values from the interval \eqn{[0,1]} indicating
the truth values. The resulting matrix has \code{nrow(x)} rows and
\code{length(condition)} columns. That is, a value on \emph{i}-th row and \emph{j}-th
column corresponds to a truth value of \emph{j}-th condition evaluated at
\emph{i}-th data row.
}
\description{
Given a data frame or a matrix of truth values of predicates, compute
truth values of given vector of conditions.
}
\details{
Each element of \code{condition} is a character string of the format \code{"{p1,p2,p3}"},
where \code{"p1"}, \code{"p2"}, and \code{"p3"} are predicates. Data \code{x} must contain columns
whose names correspond to all predicates used in conditions. Each condition
is evaluated on all data rows as an elementary conjunction, where the conjunction
operation is specified by the \code{t_norm} argument. An empty condition, \code{{}},
is always evaluated as 1.
}
\examples{
d <- data.frame(a = c(  1, 0.8, 0.5, 0.2,   0),
                b = c(0.5,   1, 0.5,   0,   1),
                c = c(0.9, 0.9, 0.1, 0.8, 0.7))
fire(d, c("{a,c}", "{}", "{a,b,c}"))
}
\author{
Michal Burda
}
