% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_condition.R
\name{format_condition}
\alias{format_condition}
\title{Format a vector of predicates into a string with a condition}
\usage{
format_condition(condition)
}
\arguments{
\item{condition}{a character vector of predicates to be formatted}
}
\value{
a character scalar with a formatted condition
}
\description{
Function takes a character vector of predicates and returns a formatted
condition. The format of the condition is a string with predicates
separated by commas and enclosed in curly braces.
}
\examples{
format_condition(NULL)              # returns {}
format_condition(c("a", "b", "c"))  # returns {a,b,c}
}
\author{
Michal Burda
}
