% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nugget.R
\name{is_nugget}
\alias{is_nugget}
\title{Test if the object is a nugget}
\usage{
is_nugget(x, flavour = NULL)
}
\arguments{
\item{x}{an object to be tested}

\item{flavour}{a character string specifying the flavour of the nugget.}
}
\value{
A logical value indicating whether the object is a nugget.
}
\description{
This function returns TRUE if the object is a nugget. If flavour is
specified, the function returns TRUE only if the object is a nugget of
the specified flavour.
}
\author{
Michal Burda
}
