% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_subset.R
\name{is_subset}
\alias{is_subset}
\title{Determine whether the first vector is a subset of the second vector}
\usage{
is_subset(x, y)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}
}
\value{
TRUE if \code{x} is a subset of \code{y} or FALSE otherwise.
}
\description{
Determine whether the first vector is a subset of the second vector
}
\author{
Michal Burda
}
