% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nugget.R
\name{nugget}
\alias{nugget}
\title{Create a nugget object of given flavour}
\usage{
nugget(x, flavour, call_function, call_args)
}
\arguments{
\item{x}{an object with rules, typically a tibble or a data frame.}

\item{flavour}{a character string specifying the flavour of the nugget or NULL.}

\item{call_function}{a function name that created the nugget.}

\item{call_args}{a list of function arguments that were used to create the nugget.}
}
\value{
A nugget object of the specified flavour, i.e., a tibble object that is
a subclass of \code{flavour} and \code{nugget} classes.
}
\description{
Create a nugget object of given flavour
}
\seealso{
\code{\link[=is_nugget]{is_nugget()}}
}
\author{
Michal Burda
}
