% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Generate all permutations of a vector}
\usage{
permute(x)
}
\arguments{
\item{x}{A vector of elements to permute. The elements can be of any type, but
they should be unique for meaningful permutations.}
}
\value{
A matrix where each row is a unique permutation of the input vector \code{x}.
}
\description{
This function generates all possible permutations of the elements in a given
vector \code{x}. The result is returned as a matrix, where each row represents a
unique permutation of the input vector.
}
\examples{
permute(c(1, 2, 3))
permute(c("a", "b", "c"))

}
\seealso{
\code{\link[utils:combn]{utils::combn()}} for combinations of elements.
}
\author{
Michal Burda
}
