% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contingency-data.frame.R,
%   R/plot_contingency-default.R, R/plot_contingency.R
\name{plot_contingency.data.frame}
\alias{plot_contingency.data.frame}
\alias{plot_contingency.default}
\alias{plot_contingency}
\title{Plot a mosaic plot for a contingency table}
\usage{
\method{plot_contingency}{data.frame}(x, ...)

\method{plot_contingency}{default}(pp, pn, np, nn, ...)

plot_contingency(x, ...)
}
\arguments{
\item{x}{A data frame with exactly one row and columns named \code{pp},
\code{pn}, \code{np}, and \code{nn}, representing the counts of true positives, false
positives, false negatives, and true negatives, respectively. All values must
be greater or equal to zero.}

\item{...}{Additional arguments (currently ignored).}

\item{pp}{The count of true positives (antecedent and consequent both true).
Value must be greater or equal to zero.}

\item{pn}{The count of false positives (antecedent true, consequent false).
Value must be greater or equal to zero.}

\item{np}{The count of false negatives (antecedent false, consequent true).
Value must be greater or equal to zero.}

\item{nn}{The count of true negatives (antecedent and consequent both false).
Value must be greater or equal to zero.}
}
\value{
A ggplot object representing the mosaic plot of the contingency table.

A ggplot object representing the mosaic plot of the contingency table.
}
\description{
Plot a mosaic plot for a contingency table

Plot a mosaic plot for a contingency table
}
\author{
Michal Burda
}
