% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Extract values from predicates}
\usage{
values(x)
}
\arguments{
\item{x}{A character vector of predicate names.}
}
\value{
A \verb{<value>} part of predicate names in \code{x}.
}
\description{
The function assumes that \code{x} is a vector of predicate names, i.e., a character
vector with elements compatible with pattern \verb{<varname>=<value>}. The function
returns the \verb{<value>} part of these elements. If the string does not
correspond to the pattern \verb{<varname>=<value>}, i.e., if the equal sign (\code{=})
is missing in the string, an empty string is returned.
}
\examples{
values(c("a=1", "a=2", "b=x", "b=y")) # returns c("1", "2", "x", "y")
}
\seealso{
\code{\link[=var_names]{var_names()}}
}
\author{
Michal Burda
}
