% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_incomparable.R
\name{which_incomparable}
\alias{which_incomparable}
\title{Return indices of first elements in a list, which are incomparable with preceding
elements.}
\usage{
which_incomparable(x, comparison)
}
\arguments{
\item{x}{a list of elements}

\item{comparison}{a binary comparison function that must return a scalar logical
value, i.e. \code{TRUE} or \code{FALSE}}
}
\value{
an integer vector of indices of selected elements.
}
\description{
The function returns indices of elements from the given list \code{x}, for which the
given binary function \code{comparison} returns \code{FALSE} for all previously selected
elements.
}
\details{
In other words, the function returns indices of selected elemnts from \code{x}. The
first element is always selected. The next element is selected only if all calls
of \code{comparable(a, b)} return \code{FALSE}, where \code{a} is the currently considered element
and \code{b} are all previously selected elements.
}
\author{
Michal Burda
}
