% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-associations.R
\name{calculate.associations}
\alias{calculate.associations}
\title{Calculate additional interest measures for association rules}
\usage{
\method{calculate}{associations}(x, measure, ...)
}
\arguments{
\item{x}{A nugget of flavour \code{associations}, typically created with
\code{\link[=dig_associations]{dig_associations()}} with argument \code{contingency_table = TRUE}.}

\item{measure}{A character vector specifying which interest measures to
calculate. See the Details section for the list of supported measures.}

\item{...}{Currently unused.}
}
\value{
An S3 object which is an instance of \code{associations} and \code{nugget}
classes and which is a tibble containing all the columns of the input
nugget \code{x}, plus additional columns for each of the requested interest
measures.
}
\description{
This function calculates various additional interest measures for
association rules based on their contingency table counts.
}
\details{
The input nugget object must contain the columns
\code{pp} (positive antecedent & positive consequent),
\code{pn} (positive antecedent & negative consequent),
\code{np} (negative antecedent & positive consequent), and
\code{nn} (negative antecedent & negative consequent), representing the counts
from the contingency table. These columns are typically produced by
\code{\link[=dig_associations]{dig_associations()}} when the \code{contingency_table} argument is set to \code{TRUE}.

The supported interest measures that can be calculated include:
\itemize{
\item \code{added_value} - see \url{https://mhahsler.github.io/arules/docs/measures#addedvalue} for details
\item \code{casual_confidence} - see \url{https://mhahsler.github.io/arules/docs/measures#casualconfidence} for details
\item \code{casual_support} - see \url{https://mhahsler.github.io/arules/docs/measures#casualsupport} for details
\item \code{centered_confidence} - see \url{https://mhahsler.github.io/arules/docs/measures#centeredconfidence} for details
\item \code{certainty} - see \url{https://mhahsler.github.io/arules/docs/measures#certainty} for details
\item \code{collective_strength} - see \url{https://mhahsler.github.io/arules/docs/measures#collectivestrength} for details
\item \code{confirmed_confidence} - see \url{https://mhahsler.github.io/arules/docs/measures#confirmedconfidence} for details
\item \code{conviction} - see \url{https://mhahsler.github.io/arules/docs/measures#conviction} for details
\item \code{cosine} - see \url{https://mhahsler.github.io/arules/docs/measures#cosine} for details
\item \code{counterexample} - see \url{https://mhahsler.github.io/arules/docs/measures#counterexample} for details
\item \code{doc} - see \url{https://mhahsler.github.io/arules/docs/measures#doc} for details
\item \code{gini} - see \url{https://mhahsler.github.io/arules/docs/measures#gini} for details
\item \code{imbalance} - see \url{https://mhahsler.github.io/arules/docs/measures#imbalance} for details
\item \code{implication_index} - see \url{https://mhahsler.github.io/arules/docs/measures#implicationindex} for details
\item \code{importance} - see \url{https://mhahsler.github.io/arules/docs/measures#importance} for details
\item \code{j_measure} - see \url{https://mhahsler.github.io/arules/docs/measures#jmeasure} for details
\item \code{jaccard} - see \url{https://mhahsler.github.io/arules/docs/measures#jaccard} for details
\item \code{kappa} - see \url{https://mhahsler.github.io/arules/docs/measures#kappa} for details
\item \code{kulczynski} - see \url{https://mhahsler.github.io/arules/docs/measures#kulczynski} for details
\item \code{lambda} - see \url{https://mhahsler.github.io/arules/docs/measures#lambda} for details
\item \code{least_contradiction} - see \url{https://mhahsler.github.io/arules/docs/measures#leastcontradiction} for details
\item \code{lerman} - see \url{https://mhahsler.github.io/arules/docs/measures#lerman} for details
\item \code{leverage} - see \url{https://mhahsler.github.io/arules/docs/measures#leverage} for details
\item \code{maxconfidence} - see \url{https://mhahsler.github.io/arules/docs/measures#maxconfidence} for details
\item \code{mutual_information} - see \url{https://mhahsler.github.io/arules/docs/measures#mutualinformation} for details
\item \code{odds_ratio} - see \url{https://mhahsler.github.io/arules/docs/measures#oddsratio} for details
\item \code{phi} - see \url{https://mhahsler.github.io/arules/docs/measures#phi} for details
\item \code{ralambondrainy} - see \url{https://mhahsler.github.io/arules/docs/measures#ralambondrainy} for details
\item \code{relative_risk} - see \url{https://mhahsler.github.io/arules/docs/measures#relativerisk} for details
\item \code{rule_power_factor} - see \url{https://mhahsler.github.io/arules/docs/measures#rulepowerfactor} for details
\item \code{sebag} - see \url{https://mhahsler.github.io/arules/docs/measures#sebag} for details
\item \code{varying_liaison} - see \url{https://mhahsler.github.io/arules/docs/measures#varyingliaison} for details
\item \code{yule_q} - see \url{https://mhahsler.github.io/arules/docs/measures#yuleq} for details
\item \code{yule_y} - see \url{https://mhahsler.github.io/arules/docs/measures#yuley} for details
}
}
\examples{
d <- partition(mtcars, .breaks = 2)
rules <- dig_associations(d,
                          antecedent = !starts_with("mpg"),
                          consequent = starts_with("mpg"),
                          min_support = 0.3,
                          min_confidence = 0.8,
                          contingency_table = TRUE)
rules <- calculate(rules,
                   measure = c("conviction", "leverage", "jaccard"))
}
\seealso{
\code{\link[=dig_associations]{dig_associations()}}
}
\author{
Michal Burda
}
