% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_condition.R
\name{format_condition}
\alias{format_condition}
\title{Format condition - convert a character vector to character scalar}
\usage{
format_condition(condition)
}
\arguments{
\item{condition}{a character vector}
}
\value{
a character scalar
}
\description{
Function takes a character vector of predicates and returns a formatted condition.
}
\author{
Michal Burda
}
