% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_diamond.R
\name{geom_diamond}
\alias{geom_diamond}
\title{Geom for drawing diamond plot representing lattice structures}
\usage{
geom_diamond(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  linetype = "solid",
  nudge_x = 0,
  nudge_y = 0.125,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings, usually created with \code{aes()}.}

\item{data}{A data frame containing the data to be plotted.}

\item{stat}{The statistical transformation to use on the data, default is "identity".}

\item{position}{Position adjustment, default is "identity".}

\item{na.rm}{A logical value indicating whether to remove missing values.}

\item{linetype}{The type of line to use for the edges, default is "solid".}

\item{nudge_x}{Horizontal nudge for the x-coordinates of the labels.}

\item{nudge_y}{Vertical nudge for the y-coordinates of the labels.}

\item{show.legend}{A logical value indicating whether to show the legend.}

\item{inherit.aes}{A logical value indicating whether to inherit aesthetics from the plot.}

\item{...}{Other arguments passed to the layer.}
}
\value{
A ggplot2 layer object that can be added to a ggplot.
}
\description{
This function creates a ggplot2 geom for drawing diamond plots, which are
used to visualize lattice structures in data. This is useful especially for
visualization of rules and its ancestors.
}
\details{
The diamonds represent items in the lattice, and edges represent
an inclusion relationship between them.
}
