% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_almost_constant.R
\name{is_almost_constant}
\alias{is_almost_constant}
\title{Tests if almost all values in a vector are the same.}
\usage{
is_almost_constant(x, threshold = 1, na_rm = FALSE)
}
\arguments{
\item{x}{a vector to be tested}

\item{threshold}{a double scalar in the interval \eqn{[0,1]} that specifies
the threshold for the proportion of the most frequent value}

\item{na_rm}{a flag indicating whether to remove \code{NA} values before testing
the proportion of the most frequent value. That is, if \code{na_rm} is \code{TRUE},
the proportion is calculated from non-\code{NA} values only. If \code{na_rm} is
\code{FALSE}, the proportion is calculated from all values and the value \code{NA}
is considered as a normal value (i.e., too much \code{NA}s can make the vector
almost constant too).}
}
\value{
If \code{x} is empty or has only one value, the function returns \code{TRUE}.
If \code{x} contains only \code{NA} values, the function returns \code{TRUE}.
If the proportion of the most frequent value is greater or equal to the
\code{threshold} argument, the function returns \code{TRUE}. Otherwise, the function
returns \code{FALSE}.
}
\description{
Function tests if almost all values in a vector are the same. The function
returns \code{TRUE} if the proportion of the most frequent value is greater or
equal to the \code{threshold} argument.
}
\examples{
is_almost_constant(1)
is_almost_constant(1:10)
is_almost_constant(c(NA, NA, NA), na_rm = TRUE)
is_almost_constant(c(NA, NA, NA), na_rm = FALSE)
is_almost_constant(c(NA, NA, NA, 1, 2), threshold = 0.5, na_rm = FALSE)
is_almost_constant(c(NA, NA, NA, 1, 2), threshold = 0.5, na_rm = TRUE)
}
\author{
Michal Burda
}
