% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuggets-package.R
\docType{package}
\name{nuggets-package}
\alias{nuggets}
\alias{nuggets-package}
\title{nuggets: Fast and Extensible Pattern Discovery in Tabular Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast and extensible framework for discovering interesting patterns in tabular data. The package searches combinations of fuzzy or Boolean predicates and evaluates the resulting subgroups using statistical, logical, or structural measures. It supports a broad range of pattern-discovery tasks, including association rules (Agrawal et al., 1994, \url{https://www.vldb.org/conf/1994/P487.PDF}), contrast patterns (Chen, 2022, \doi{10.48550/arXiv.2209.13556}), emerging patterns (Dong et al., 1999, \doi{10.1145/312129.312191}), subgroup discovery (Atzmueller, 2015, \doi{10.1002/widm.1144}), and conditional correlations (Hájek, 1978, \doi{10.1007/978-3-642-66943-9}). User-defined functions may be supplied to guide custom pattern searches, making the framework applicable beyond traditional assoociation-rule mining. Efficient implementation enables pattern discovery on large and dense data sets. Package includes methods for visualization and supports interactive exploration through integrated Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beerda.github.io/nuggets/}
  \item \url{https://github.com/beerda/nuggets}
  \item Report bugs at \url{https://github.com/beerda/nuggets/issues}
}

}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

Authors:
\itemize{
  \item Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})
}

}
\keyword{internal}
