% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuggets-package.R
\docType{package}
\name{nuggets-package}
\alias{nuggets}
\alias{nuggets-package}
\title{nuggets: Extensible Framework for Data Pattern Exploration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A framework for systematic exploration of association rules (Agrawal et al., 1994, \url{https://www.vldb.org/conf/1994/P487.PDF}), contrast patterns (Chen, 2022, \doi{10.48550/arXiv.2209.13556}), emerging patterns (Dong et al., 1999, \doi{10.1145/312129.312191}), subgroup discovery (Atzmueller, 2015, \doi{10.1002/widm.1144}), and conditional correlations (Hájek, 1978, \doi{10.1007/978-3-642-66943-9}). User-defined functions may also be supplied to guide custom pattern searches. Supports both crisp (Boolean) and fuzzy data. Generates candidate conditions expressed as elementary conjunctions, evaluates them on a dataset, and inspects the induced sub-data for statistical, logical, or structural properties such as associations, correlations, or contrasts. Includes methods for visualization of logical structures and supports interactive exploration through integrated Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beerda.github.io/nuggets/}
  \item \url{https://github.com/beerda/nuggets/}
  \item Report bugs at \url{https://github.com/beerda/nuggets/issues/}
}

}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

}
\keyword{internal}
