% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuggets-package.R
\docType{package}
\name{nuggets-package}
\alias{nuggets}
\alias{nuggets-package}
\title{nuggets: Extensible Data Pattern Searching Framework}
\description{
Generate conditions in the form of elementary conjunctions, evaluate them on a dataset and check the induced sub-data for interesting statistical properties. Currently, the package searches for implicative association rules and conditional correlations. A user-defined function may be defined to evaluate on each generated condition to search for custom patterns.
}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

}
\keyword{internal}
