% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuggets-package.R
\docType{package}
\name{nuggets-package}
\alias{nuggets}
\alias{nuggets-package}
\title{nuggets: Nuggets}
\description{
A package for searching interesting subspaces described with elementary conjunctions.
}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

}
\keyword{internal}
