% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuggets-package.R
\docType{package}
\name{nuggets-package}
\alias{nuggets}
\alias{nuggets-package}
\title{nuggets: Extensible Data Pattern Searching Framework}
\description{
Extensible framework for subgroup discovery (Atzmueller (2015) \doi{10.1002/widm.1144}), contrast patterns (Chen (2022) \doi{10.48550/arXiv.2209.13556}), emerging patterns (Dong (1999) \doi{10.1145/312129.312191}) and association rules (Agrawal (1994) \url{https://www.vldb.org/conf/1994/P487.PDF}). Both crisp (binary) and fuzzy data are supported. It generates conditions in the form of elementary conjunctions, evaluates them on a dataset and checks the induced sub-data for interesting statistical properties. Currently, the package searches for implicative association rules and conditional correlations (Hájek (1978) \doi{10.1007/978-3-642-66943-9}). A user-defined function may be defined to evaluate on each generated condition to search for custom patterns.
}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

}
\keyword{internal}
