% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_condition.R
\name{parse_condition}
\alias{parse_condition}
\title{Convert character vector of conditions into a list of vectors of predicates}
\usage{
parse_condition(..., .sort = FALSE)
}
\arguments{
\item{...}{character vectors of conditions to be parsed.}

\item{.sort}{a flag indicating whether to sort the predicates in the result.}
}
\value{
a list of vectors of predicates with each element corresponding to one
condition.
}
\description{
Function takes a character vector of conditions and returns a list of vectors
of predicates. Each element of the list corresponds to one condition. The
condition is a string with predicates separated by commas and enclosed in
curly braces, as returned by \code{\link[=format_condition]{format_condition()}}. The function splits the
condition string into a vector of predicates.
}
\details{
If multiple vectors of conditions are passed, each of them is processed
separately and the result is merged into a single list element-wisely. If
the lengths of the vectors are different, the shorter vectors are recycled.
}
\examples{
parse_condition(c("{a}", "{x=1, z=2, y=3}", "{}"))
parse_condition(c("{b}", "{x=1, z=2, y=3}", "{q}", "{}"),
                c("{a}", "{v=10, w=11}",    "{}",  "{r,s,t}"))

}
\author{
Michal Burda
}
