% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contingency-data.frame.R,
%   R/plot_contingency-default.R, R/plot_contingency.R
\name{plot_contingency.data.frame}
\alias{plot_contingency.data.frame}
\alias{plot_contingency.default}
\alias{plot_contingency}
\title{Plot a mosaic plot for a contingency table}
\usage{
\method{plot_contingency}{data.frame}(d, ...)

\method{plot_contingency}{default}(pp, pn, np, nn, ...)

plot_contingency(...)
}
\arguments{
\item{d}{A data frame with exactly one row and columns named \code{pp},
\code{pn}, \code{np}, and \code{nn}, representing the counts of true positives, false
positives, false negatives, and true negatives, respectively. All values must
be greater or equal to zero.}

\item{...}{Additional arguments (currently ignored).}

\item{pp}{The count of true positives (antecedent and consequent both true).
Value must be greater or equal to zero.}

\item{pn}{The count of false positives (antecedent true, consequent false).
Value must be greater or equal to zero.}

\item{np}{The count of false negatives (antecedent false, consequent true).
Value must be greater or equal to zero.}

\item{nn}{The count of true negatives (antecedent and consequent both false).
Value must be greater or equal to zero.}
}
\value{
A ggplot object representing the mosaic plot of the contingency table.
}
\description{
This function creates a mosaic plot for a contingency table defined by the
counts of true positives, false positives, false negatives, and true negatives.
The plot visually represents the distribution of these counts in a 2x2 grid.
The area of each rectangle in the plot corresponds to the count of the respective category.
Vertical and horizontal lines are added to the plot to indicate the expected
proportions of the counts under the assumption of independence between the antecedent
and the consequent.
}
\examples{
plot_contingency(pp = 30, pn = 10, np = 20, nn = 40)
}
\author{
Michal Burda
}
