% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ill_conditions.R
\name{remove_ill_conditions}
\alias{remove_ill_conditions}
\title{From a given list remove those elements that are not valid conditions.}
\usage{
remove_ill_conditions(x, data)
}
\arguments{
\item{x}{a list of character vector}

\item{data}{a matrix or a data frame}
}
\value{
a list of elements of \code{x} that are valid conditions.
}
\description{
A valid condition is a character vector of predicates, where each predicate
corresponds to some column name of the related data frame.
(An empty character vector is considered as a valid condition too.)
}
\seealso{
\code{\link[=is_condition]{is_condition()}}
}
\author{
Michal Burda
}
